chore: move swagger docs to separate package
