+++ controller/user.go
@@ -2,11 +2,15 @@ package controller
 
 import (
 	"go-api-template/model/commonerrors"
+	"go-api-template/model/swagger"
 	"go-api-template/service"
 
 	"github.com/gin-gonic/gin"
 	"github.com/go-openapi/strfmt"
 )
+
+// To avoid unused dependency
+var _ = swagger.BadRequestResponse{}
 
 type IUser interface {
 	UserByID(ctx *gin.Context)
@@ -30,10 +34,10 @@ func NewUser(service service.IUser) IUser {
 //	@Accept			json
 //	@Produce		json
 //	@Param			user_id	path		string							true	"User ID"	format(uuid)
-//	@Success		200		{object}	model.UserByIDResponseSwagger	"Get user by ID"
-//	@Failure		400		{object}	model.BadRequestResponse
-//	@Failure		403		{object}	model.ForbiddenResponse
-//	@Failure		500		{object}	model.InternalErrorResponse
+//	@Success		200		{object}	swagger.UserByIDResponseSwagger	"Get user by ID"
+//	@Failure		400		{object}	swagger.BadRequestResponse
+//	@Failure		403		{object}	swagger.ForbiddenResponse
+//	@Failure		500		{object}	swagger.InternalErrorResponse
 //	@Router			/users/{user_id} [get]
 func (controller *user) UserByID(ctx *gin.Context) {
 	// Validate path params
