+++ model/swagger/created_response.go
@@ -1,4 +1,4 @@
-package model
+package swagger
 
 type CreatedResponse struct {
 	Message string `json:"message" example:"Resource created successfully"`
