+++ model/swagger/forbidden_response.go
@@ -1,4 +1,4 @@
-package model
+package swagger
 
 type ForbiddenResponse struct {
 	Message string `json:"message" example:"You don't have permission to access this resource"`
