+++ model/swagger/ok_response.go
@@ -1,4 +1,4 @@
-package model
+package swagger
 
 type OKResponse struct {
 	Message string `json:"message" example:"Success"`
