+++ model/swagger/unauthorized_response.go
@@ -1,4 +1,4 @@
-package model
+package swagger
 
 type UnauthorizedResponse struct {
 	Message string `json:"message" example:"You are not authorized to access this resource"`
