+++ model/swagger/user_by_id_response_swagger.go
@@ -1,4 +1,4 @@
-package model
+package swagger
 
 type UserByIDResponseSwagger struct {
 	User *UserSwagger `json:"user"`
