+++ model/swagger/user_swagger.go
@@ -1,4 +1,4 @@
-package model
+package swagger
 
 import "time"
